Refactor project structure to separate core logic and CLI
This commit refactors the projects in the `projects/` directory to enforce a consistent structure where business logic is separated from the user interface.

For the `pointsys` project:
- All core logic was extracted from `main.py` into a new `projects/pointsys/core.py`.
- A new command-line interface was created in `projects/pointsys/cli.py`.
- The original `projects/pointsys/main.py` was removed.

For the `tasklist` project:
- A new `projects/tasklist/cli.py` was created to provide a command-line interface for the existing `core.py` and `data.py` modules.
